correções solicitadas conforme analise do coderewiew solcitado

+++ src/main/java/com/clientes/cadastro/service/ClientService.java
@@ -1,8 +1,8 @@
 package com.clientes.cadastro.service;
 
+import com.clientes.cadastro.exception.BadRequest;
 import com.clientes.cadastro.model.Client;
 import com.clientes.cadastro.repository.ClientRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,10 +11,23 @@ import java.util.Optional;
 @Service
 public class ClientService {
 
-    @Autowired
     private ClientRepository clientRepository;
+    public ClientService(ClientRepository clientRepository){
+        this.clientRepository = clientRepository;
+    }
 
-    public Client registerClient(Client client) {
+    //comportamentos
+    public Client registerClient(Client client) throws Exception{
+        Optional<Client> nameClientOpt = clientRepository.findByEmail(client.getEmail());
+        Optional<Client> emailClientOpt = clientRepository.findByEmail(client.getEmail());
+        Optional<Client> ssrClientOpt = clientRepository.findByDocument(client.getDocument());
+
+        if(emailClientOpt.isPresent()){
+            throw new BadRequest("Email already exists");
+        }
+        if(ssrClientOpt.isPresent()){
+            throw new BadRequest("CPF already exists");
+        }
         return clientRepository.save(client);
     }
 
@@ -23,25 +36,28 @@ public class ClientService {
     }
 
     public Optional<Client> getOnlyOneClient(Long id) {
+
+        Optional<Client> findClient = clientRepository.findById(id);
+        if(!findClient.isPresent()){
+            throw new BadRequest("Cliente com o Id " + id + " não foi encontrado.");
+        }
         return clientRepository.findById(id);
     }
 
-    public Optional<Client> getByName(String name) {
-        return clientRepository.findByName(name);
+    public List<Client> searchByNameParts(String name) {
+        return clientRepository.findByNameContainingIgnoreCase(name);
     }
 
-    public Client registeredClient(Client client) {
+    public Client updateClientId(Client client) {
         return clientRepository.save(client);
     }
 
     public void deleteCustomerRecord(Long id) {
+        Optional<Client> findClient = clientRepository.findById(id);
+        if(!findClient.isPresent()){
+            throw new BadRequest("Cliente com o Id " + id + " não foi encontrado.");
+        }
         clientRepository.deleteById(id);
     }
 
-
 }
-
-/* ESTUDAR O CASO: >> com pesquisar por nome que pode me trzer uma lista porque tem várias pessoas com um nome em comum???*/
-//public List<Client> getByName(String name) {
-//    return clientRepository.findByName(name);
-//}
